Reorganizing strategy.transaction as a submodule.
TransactionStrategy has been renamed to TransactionTraversal.
A new Strategy subclass wrapping TransactionTraversal will be added next
and will use the TransactionStrategy name for consistency.
This change touches a lot of files.

TransactionNode now has its own module ("node").
TransactionTraversal moved to new module ("base").
Other methods moved to "util" module.
Empty "strategy" module added as placeholder.

Placeholder testing modules created for new submodules of
strategy.transaction (i.e. those other than base)
