+++ forecaster/strategy/transaction/strategy.py
@@ -1,0 +1,7 @@
+""" Provides Strategy-type wrappers for TransactionTraversal. """
+
+from forecaster.strategy.base import Strategy
+
+class TransactionStrategy(Strategy):
+    """ TODO """
+    pass
