+++ tests/strategy/transaction/test_strategy.py
@@ -1,0 +1,3 @@
+""" TODO """
+
+# TODO: Add tests for TransactionStrategy.
